requirements.txt
Django>=3.2.4,<4.0
djangorestframework>=3.14
psycopg2-binary>=2.9
redis>=5.0.0
python-json-logger>=2.0.7,<2.1
drf-yasg>=1.21
